Clarify objects' responsibility
Curses windows are now initialized at the top of the application,
and passed as an argument, instead of in each object individually.
As a result, a window is shared by all tabs.
This must contribute to smaller heap allocation.

Singleton objects should generally be avoided.

There are still awkward dependencies between objects, which need to be
resolved someday.